+++ config/webpack/development.js
@@ -1,14 +1,11 @@
 process.env.NODE_ENV = process.env.NODE_ENV || 'development';
 
 const config = require('./shared');
-const eslint = require('./loaders/eslint');
-
-config.module.rules.unshift(eslint);
 
 config.devServer = {
   ...config.devServer,
   clientLogLevel: 'debug',
-  writeToDisk: true
-}
+  writeToDisk:    true
+};
 
 module.exports = config;