+++ config/webpack/shared.js
@@ -6,6 +6,8 @@ const yaml = require('./loaders/yaml');
 const I18n = require('./loaders/I18n');
 const modulejs = require('./loaders/modulejs');
 const tooltip = require('./loaders/tooltip');
+const eslint = require('./loaders/eslint');
+const ts = require('./loaders/ts');
 const SymlinkAssets = require('./plugins/symlink');
 
 environment.plugins.prepend('Provide', new webpack.ProvidePlugin({
@@ -20,6 +22,8 @@ environment.plugins.prepend('clean', new CleanWebpackPlugin());
 
 environment.loaders.append('yaml', yaml);
 
+environment.loaders.append('eslint', eslint);
+environment.loaders.append('ts-loader', ts);
 // Expose libs called from haml templates
 environment.loaders.append('modulejs', modulejs);
 environment.loaders.append('i18n-js', I18n);