+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+v2.7.0 [unreleased]
+-------------------
+### Features
+- [#562](https://github.com/latera/homs/pull/562) Bring in TypeScript into the project.
+
 v2.6.2 [unreleased]
 -------------------
 ### Features